Move memory CLI front door to generated entrypoint
