+++ packages/command-generation/src/agentic_command_generation/generator.py
@@ -45,7 +45,7 @@ def _runtime_command_for_package(package: dict[str, Any], runtime_binding: dict[
     if package_role == "planning-module-cli":
         return "python -m repo_planning_bootstrap.generated_cli_entrypoint"
     if package_role == "memory-module-cli":
-        return "python -m repo_memory_bootstrap.cli"
+        return "python -m repo_memory_bootstrap.generated_cli_entrypoint"
     return str(runtime_binding["default_runtime_command"])
