+++ packages/memory/pyproject.toml
@@ -30,7 +30,7 @@ Repository = "https://github.com/rickardvh/agentic-workspace"
 Issues = "https://github.com/rickardvh/agentic-workspace/issues"
 
 [project.scripts]
-agentic-memory = "repo_memory_bootstrap.cli:main"
+agentic-memory = "repo_memory_bootstrap.generated_cli_entrypoint:main"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/repo_memory_bootstrap"]
@@ -81,4 +81,3 @@ source = ["repo_memory_bootstrap"]
 [tool.coverage.report]
 show_missing = true
 skip_covered = true
-
