+++ packages/memory/src/repo_memory_bootstrap/__main__.py
@@ -1,4 +1,4 @@
-from repo_memory_bootstrap.cli import main
+from repo_memory_bootstrap.generated_cli_entrypoint import main
 
 if __name__ == "__main__":
     raise SystemExit(main())
