+++ src/agentic_workspace/contracts/proof_selection_rules.json
@@ -197,7 +197,7 @@
         "role": "hand-owned-executable",
         "exact": [
           "packages/planning/src/repo_planning_bootstrap/_runtime_cli.py",
-          "packages/memory/src/repo_memory_bootstrap/cli.py"
+          "packages/memory/src/repo_memory_bootstrap/_runtime_cli.py"
         ],
         "direct_edit_allowed": true,
         "source_contract": "package-local CLI source plus package bootstrap payload checks",
